Fri, Jul 21, 2023, 9:04 PM +01:00

+++ professor/check.py
@@ -9,7 +9,7 @@ def main():
 
 def check(s):
     try:
-        t = float(s)
+        t = int(s)
         return t
     except ValueError:
         pass
